requirements.txt
Django==4.2
djangorestframework==3.15.2
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
Faker==30.8.1
drf-yasg==1.21.8